+++ HISTORY.rst
@@ -7,6 +7,7 @@ History
     * New task to notify professionals about Assignments with late submissions (rudaporto).
     * New task to notify professionals before shooting (rudaporto).
     * Card #451: On Order export use the vocabulary value of Asset Type (ericof).
+    * Card #468: On Assignment cancellation set travel expenses to zero (ericof).
 
 2.1.27 (2017-07-12)
 -------------------